requirements.txt
streamlit==1.38.0
pandas==2.2.2
gspread==6.0.0
google-auth==2.34.0
